change order in head. add css link. add default content containers.
Updated to make the page a bit more usable in a normal web site. removed the class="no-js" in the html element as this is only needed for something like Modernizr. changed order in head to group things. Added a link to a stylesheet as most sites have one. Added in head/main/footer elements in the body as this is also typical of most sites.